+++ packages/auth/src/index.ts
@@ -1,4 +1,11 @@
-export * from "./auth";
+export {
+    authenticateInteractively,
+    authenticateSilently,
+    authenticationClientExists,
+    initiateAuthenticationClient,
+    getAccount,
+    signOut,
+} from "./auth";
 export * from "./types";
 export * from "./components";
 export * from "./hooks";